pyproject.toml
[tool.poetry]
name = "sentri-interview"
version = "0.1.0"
description = ""
authors = ["Huayuan Tu <tuhuayuan@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
pyjwt = "^2.8.0"
apscheduler = "^3.10.4"
redis = "^5.0.6"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
httpx = "^0.27.0"
ipython = "^8.25.0"
mypy = "^1.10.0"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
